Transformer: move Transformer related classes into onmt/modules/Transformer.py
1) Move TransformerDecoder and TransformerDecoderState to
   onmt/modules/Transformer.py.
2) Rename the original TransformerDecoder to TransformerDecoderLayer.
3) Move DecoderState to onmt/Translator.py, otherwise we will circular dependency
   problem between onmt/Models.py and onmt/modules/Transformer.py.
